requirements.txt
aiobotocore[boto3,awscli]
django-phonenumber-field[phonenumbers]
Django==3.2.12
openpyxl==3.0.7
pandas==1.3.5
psycopg2==2.9.3
python-slugify==5.0.2
pytz
requests
s3fs~=0.4.0
twilio
